email-print-mime-structure: change --use-gpg-agent to a simple flag
Turns out that type=bool doesn't really do what we want it to do (see
https://bugs.python.org/issue37564), and there is no built-in easy
answer for argparse to accept a boolean value sensibly
(e.g. type='bool', which might be able to handle "yes" and "no" and
"1" and "0" and "on" and "off" as well as "true" and "false", etc)

So rather than implement all of that here, we'll just have
--use-gpg-agent as a simple flag.  This is an API change, but the
previous API has only been out for a few days, and the tool is
documented for interactive use.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>