Add imap-dl, a simple imap downloader
getmail upstream appears to have no plans to convert to python3 in the
near future.

Some of us use only a minimal subset of features of getmail, and it
would be nice to have something simpler, with the main complexity
offloaded to the modern python3 stdlib.

This patch represents a squashed series of changes from both Jameson
Graef Rollins and Daniel Kahn Gillmor (dkg), though dkg is primarily
responsible for any remaining bugs.

Thanks to Sean Whitton for useful and significant feedback.

Signed-off-by: Jameson Graef Rollins <jrollins@finestructure.net>
Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>