d/control: add Homepage field
If mailscripts ever grows a proper webpage that's not just a Vcs, we
can change the Homepage: field to point to it at that time.  In the
meantime, point to the most visible web resource available to the
project.

It's redundant with the Vcs-Browser: field, but this is still useful
for anyone who is searching just the Homepage: fields for information
about the projects in debian.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>