email-print-mime-structure: decryption/unwrapping uses BOX_DRAWINGS
The symbols ↧ and ⇩ were used for decryption and unwrapping,
respectively, but they are not included in very many fonts.

By switching the tree structure to only use characters from the BOX
DRAWINGS subset of Unicode, we have a better chance of producing
reliable renderings in different fonts.

See also discussion at
https://github.com/ietf-tools/xml2rfc/issues/1245#issuecomment-3014573118
which motivated this cleanup.

Note that redrawing the tree this way needs an adjustment in the line
above the "(decrypts to)" or "(unwraps to)" line, so we handle that as
well.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>